Noxi/requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
fonttools==4.39.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.27.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lime==0.2.0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
networkx==3.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.28.2
rfc3986==1.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
threadpoolctl==3.1.0
tifffile==2023.3.21
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
gunicorn==20.1.0
APScheduler~=3.10.1
nltk~=3.8.1